feat(LINGUIST-353): Moved enums package to root, IntelliJ IDE refactored imports. (#57)
