+++ src/main/java/app/linguistai/bmvp/enums/LogType.java
@@ -1,4 +1,4 @@
-package app.linguistai.bmvp.model.enums;
+package app.linguistai.bmvp.enums;
 
 public enum LogType {
     INFO("INFO"),
