+++ src/main/java/app/linguistai/bmvp/response/wordbank/RUnknownWord.java
@@ -1,6 +1,6 @@
 package app.linguistai.bmvp.response.wordbank;
 
-import app.linguistai.bmvp.model.enums.ConfidenceEnum;
+import app.linguistai.bmvp.enums.ConfidenceEnum;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
