+++ src/test/java/app/linguistai/bmvp/service/gamification/XPServiceTest.java
@@ -3,7 +3,7 @@ package app.linguistai.bmvp.service.gamification;
 import app.linguistai.bmvp.configs.XPConfiguration;
 import app.linguistai.bmvp.exception.NotFoundException;
 import app.linguistai.bmvp.model.User;
-import app.linguistai.bmvp.model.enums.XPAction;
+import app.linguistai.bmvp.enums.XPAction;
 import app.linguistai.bmvp.model.gamification.UserXP;
 import app.linguistai.bmvp.repository.IAccountRepository;
 import app.linguistai.bmvp.repository.gamification.IUserXPRepository;
