+++ src/test/java/app/linguistai/bmvp/service/wordbank/UnknownWordServiceTest.java
@@ -2,7 +2,7 @@ package app.linguistai.bmvp.service.wordbank;
 
 import app.linguistai.bmvp.exception.NotFoundException;
 import app.linguistai.bmvp.model.User;
-import app.linguistai.bmvp.model.enums.ConfidenceEnum;
+import app.linguistai.bmvp.enums.ConfidenceEnum;
 import app.linguistai.bmvp.model.wordbank.ListStats;
 import app.linguistai.bmvp.model.wordbank.UnknownWord;
 import app.linguistai.bmvp.model.wordbank.UnknownWordList;
